requirements.txt
alembic==1.13.1
async-timeout==4.0.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.2
MarkupSafe==2.1.5
nba_api==1.4.1
numpy==1.24.4
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyarrow==15.0.0
pytest==8.0.0
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
rq==1.15.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.26
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
